+++ src/main/scala/de/uni_luebeck/isp/tessla/tessladoc/Main.scala
@@ -15,6 +15,7 @@ object Main extends SexyOpt {
   val stdLib = flag("stdlib", 's', "Include documentation for definitions from the standard library")
   val includes = flag("includes", 'i', "Include documentation from included files")
   val globalsOnly = flag("globals-only", 'g', "Do not show information for local definitions")
+  val html = flag("html", "Generate documentation as HTML")
 
   def main(args: Array[String]) = {
     parse(args)
@@ -24,7 +25,8 @@ object Main extends SexyOpt {
       case Success(tesslaDocs, warnings) =>
         warnings.foreach(w => System.err.println(s"Warning: $w"))
         val relevantDocs = if(globalsOnly) tesslaDocs.globalsOnly else tesslaDocs
-        println(relevantDocs)
+        val output = if (html) HtmlGenerator.generateHTML(relevantDocs) else relevantDocs.toString
+        println(output)
       case Failure(errors, warnings) =>
         warnings.foreach(w => System.err.println(s"Warning: $w"))
         errors.foreach(e => System.err.println(s"Error: $e"))